reactor_agent/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-openai>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0
pdfplumber>=0.11.0
fpdf2>=2.7.9
reportlab>=4.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
# LLM provider SDKs (install only what you need)
groq>=0.11.0
openai>=1.0.0
anthropic>=0.34.0
ollama>=0.3.0
